+++ apps/x/packages/core/src/runtime/tools/domains/background-tasks.ts
@@ -3,12 +3,68 @@
 // catalog key order (provider-payload bytes; see the key-order test there).
 
 import { z } from "zod";
-import {
-    CreateBackgroundTaskInput,
-    PatchBackgroundTaskInput,
-    resolveCodeProject,
-    BuiltinToolsSchema,
-} from "./support.js";
+import { BuiltinToolsSchema } from "../types.js";
+
+import * as path from "path";
+import container from "../../../di/container.js";
+import { BackgroundTaskSchema, TriggersSchema } from "@x/shared/dist/background-task.js";
+import * as gitService from "../../../code-mode/git/service.js";
+import type { ICodeProjectsRepo } from "../../../code-mode/projects/repo.js";
+import { expandHome } from "../paths.js";
+
+// Inputs for the bg-task builtin tools. Reuse the canonical schema field
+// descriptions; only `triggers` gets a tighter contextual override (the
+// shared TriggersSchema description is written from the live-note perspective).
+export const CreateBackgroundTaskInput = BackgroundTaskSchema.pick({
+    name: true,
+    instructions: true,
+    triggers: true,
+    model: true,
+    provider: true,
+}).extend({
+    triggers: TriggersSchema.optional().describe('All three sub-fields (cronExpr, windows, eventMatchCriteria) are independently optional — mix freely. No triggers at all = manual-only (user clicks Run).'),
+    projectDir: z.string().optional().describe(
+        "Set this ONLY when the user wants the task to WRITE CODE. An absolute path (or ~/…) to a LOCAL GIT REPOSITORY with at least one commit. It turns this into a *coding task*: each run scans the trigger source for actionable items and implements them autonomously in isolated git worktrees off this repo — never touching the user's checkout. Extract the directory from the user's request (e.g. 'use ~/Work/space/test as the work directory'). Omit for ordinary output/action tasks.",
+    ),
+});
+
+export const PatchBackgroundTaskInput = BackgroundTaskSchema.pick({
+    name: true,
+    instructions: true,
+    active: true,
+    triggers: true,
+    model: true,
+    provider: true,
+}).partial().extend({
+    slug: z.string().describe('The slug of the task to update (the folder name under bg-tasks/).'),
+    triggers: TriggersSchema.optional().describe('Replace the triggers object. To remove all triggers (make manual-only) pass an empty object.'),
+    projectDir: z.string().optional().describe("Point an existing task at a code repo (or change which one) to make it a coding task. Absolute path or ~/… to a local git repository with at least one commit. Same rules as on create."),
+    clearModel: z.boolean().optional().describe("Reset the task's model/provider override so it falls back to the default. Use this to unstick a bad/rejected model value (do not also pass model)."),
+});
+
+export async function resolveCodeProject(dirPath: string): Promise<
+    { ok: true; projectId: string; path: string; warning?: string } | { ok: false; error: string }
+> {
+    const abs = path.resolve(expandHome(dirPath));
+    const projectsRepo = container.resolve<ICodeProjectsRepo>('codeProjectsRepo');
+    let project: Awaited<ReturnType<ICodeProjectsRepo['add']>>;
+    try {
+        project = await projectsRepo.add(abs);
+    } catch (err) {
+        return { ok: false, error: `Could not use '${dirPath}' as a code directory: ${err instanceof Error ? err.message : String(err)}` };
+    }
+    // Worktree isolation needs a real git repo with at least one commit
+    // (codeSessionService.create throws otherwise). Surface it now as a soft
+    // warning rather than letting the next run fail silently.
+    let warning: string | undefined;
+    try {
+        const info = await gitService.repoInfo(project.path);
+        if (!info.isGitRepo) warning = `${project.path} is not a git repository yet — run \`git init\` and make a commit, or the coding sessions will fail.`;
+        else if (!info.hasCommits) warning = `${project.path} has no commits yet — make an initial commit, or the coding sessions will fail.`;
+    } catch { /* best effort — worktree creation will surface it later */ }
+    return { ok: true, projectId: project.id, path: project.path, ...(warning ? { warning } : {}) };
+}
+
 
 
 export const backgroundTaskTools: z.infer<typeof BuiltinToolsSchema> = {
