+++ apps/x/packages/core/src/runtime/tools/domains/shell.ts
@@ -8,9 +8,7 @@ import { executeCommand, executeCommandAbortable } from "../../../application/li
 import { agentSlackShimEnv } from "../../../slack/agent-slack-exec.js";
 import { WorkDir } from "../../../config/config.js";
 import type { ToolContext } from "../exec-tool.js";
-import {
-    BuiltinToolsSchema,
-} from "./support.js";
+import { BuiltinToolsSchema } from "../types.js";
 
 
 export const shellTools: z.infer<typeof BuiltinToolsSchema> = {
