+++ apps/x/packages/core/src/runtime/tools/domains/web.ts
@@ -9,9 +9,7 @@ import { WorkDir } from "../../../config/config.js";
 import { isSignedIn } from "../../../account/account.js";
 import { getAccessToken } from "../../../auth/tokens.js";
 import { API_URL } from "../../../config/env.js";
-import {
-    BuiltinToolsSchema,
-} from "./support.js";
+import { BuiltinToolsSchema } from "../types.js";
 
 
 export const webSearchTools: z.infer<typeof BuiltinToolsSchema> = {
